regression_tests: ovs-dpdk-pvp add emulatorpin to the guest
This cpu will end up being used for the management of the vm, this can
improve the stability of the results. Setting the default value of the
emulator pin alias to 9 (as in core 9) as this is the next free core in
our usual setup.

Also setting the default value for guest cores to 5,6,7,8 for the same
reason.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>