regression_tests: ovs-dpdk-pvp add guest_testpmd_cores alias
The first '-c <value>' argument of testpmd specifies the coremask for
the whole testpmd application, this includes all pmd operations (2 dpdk
nics in our case) and the management thread. For optimal performance we
should therefore provide 3 cores. This will now be supplied by the
guest_testpmd_cores alias (default 0x7 - cores 0,1,2).

The second (after '-- ') '--coremask' argument specifies the coremask
for the pmd threads (2 in our case) where we'll specify whic 2 of the 3
cores will be used for the dpdk nics. This will be supplied by the
guest_dpdk_cores alias (default 0x6 - cores 1,2).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>