regression_tests: ovs-dpdk-pvp fix guest hugepages memory
For the guest to properly use hugepages it is necessary to configure
memory backing using them.

It's also important to ensure that the host machine has enough hugepages
to cover all the ram of the guest. In our setup we use a guest with 16G
ram so we need to increase the number of 2M pages to at least 8000. I'm
setting it to 13000 just to be sure.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>