regression_tests: ovs-dpdk-pvp: change guest connect logic
Instead of waiting an arbitrary 60 seconds for the guest to start then
trying a single ssh connection we should be repeatedly trying to ssh
connect until it succeeds and set an upper limit.

For several past runs the guest startup time was much more unpredictable
(maybe due to more isolated cores on the host?) so this upper time limit
should be much larger. The implemented loop should try 60 times to
connect with 5s waits in between tries. In the worst case scenario this
should be enough of an upper bound.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>