regression_tests: ovs-dpdk-pvp remove warmup pings
The warmup pings are not actually required for the setup to work, and
with the recent change to the source bind, the second ping on each host
(so nic2 -> nic2) stopped working due the addresses being in the same
network and how ARP resolution is optimized by the kernel - the ping
requests leave from nic2 but arrive on nic1 on the other host and the
reply then also leaves from that nic and returns on the nic1 of the
source host. This means that the ping utility doesn't receive the reply
message and reports a failure.

I'm keeping the configuration of the ip addresses because that's part of
the PerfRepo result parameters and removing it changes the hashes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>