lnst.Common.Utils: change std_deviation calculation
The old algorithm works and has the advantage of a single pass throught
the value array, however in case of identical small values (less than 1)
it might encounter an error of calculating the square root of a negative
number instead of properly returning 0.
Example:
[0.031, 0.031, 0.031, 0.031, 0.031]

The new algorithm is less efficient (uses 2 iterations of the value
array) but shouldn't have the same issue.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>