RecipeCommon: IRQ: check if MSI interrupt is registered
This is a fix for an issue we hit on specific NIC tracked in
https://bugzilla.redhat.com/show_bug.cgi?id=1484059

In this case the mlx4_core driver lists all available MSI interrupts
under /sys/class/$dev/device/msi_irqs, this covers interrupts for both
mlx4_en (network part) and mlx4_ib (storage part) even if any of these is
not loaded yet. The IRQs are propagated to /proc/irq/ directory once
they're loaded.

To fix this the availability of the interrupt id is first checked in
/proc/irq/ directory and modified only if present.

v2: I moved the /proc/irq listing out of the loop so that the command is
    run just once and reused when checking the presence of specific IRQ
    number

Signed-off-by: Jan Tluka <jtluka@redhat.com>