Slave: fallback to iproute commands if brctl is missing
Bridge device can be configured by iproute for quite a long time and brctl
is getting deprecated in recent distributions.

This patch replaces brctl with iproute alternative when creating a bridge
device. For bridge port addition and removal both brctl and iproute method
is provided depending on brctl availability.

Tested on RHEL6, RHEL7 and Fedora28.

Signed-off-by: Jan Tluka <jtluka@redhat.com>