lnst.Common.Parameters: add ListParam
ListParam accepts list objects (not other iterables, such as a tuple or
string), and accepts an optional type parameter, if this parameter is
provided it will be used to type check each individual item in the list.
This can be useful for occasions where you want a parameter that is a
list of Integers or Strings.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>