regression_tests: add ovs rpm version to TestExec comment
To get the version of the main ovs package while supporting future
openvswitch2.10 and other packages we use the rpm query by file option.
This isn't universal and won't work on distributions with different
package systems, so we also add || true to the command to avoid failures
because of this command.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>