lnst.Controller.RecipeResults: add data_level attribute
The level attribute specifies the importance of the Result object that
is used either for filtering or formatting purposes when processing the
recipe results.

The data_level attribute is an extension of that by specifying the
importance of the data provided with the result. It is used by the
RunSummaryFormatter to filter out the data provided with the result.

The default for the Base class is ResultLevel.DEBUG same as the level
attribute.

For the JobResult class it's always level+1, for ease of use when
formatting results:
    * choose filter level -> show results
    * choose filter level + 1 -> show results and their data

For the Result class used by user, the default is
ResultLevel.IMPORTANT+1 (so level+1 same as the JobResult), but the user
has the ability to change this when calling Recipe.add_result

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>