lnst.Controller.Namespace: add prepare_job method for delayed start
The prepare_job method will create and return a lnst.Controller.Job
object the same as the Namespace.run method but won't send the command
to start it to the Slave. Instead the tester can call the Job.start
method himself to send the start command later.

This could be used to achieve better grouping of time related job
starts, currently it would only be useful if you intend to do resource
intensive work between Namespace.run calls, but I can imagine extending
this functionality to actually provide a more intelligent synchronized
start of multiple jobs.

Consider this just an idea, might be removed later.

v2:
    * the Namespace.run method is now just a shortcut for prepare_job +
      job.start. The exception handling isn't necessary and is removed
      in this version

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>