add lnst.Tests.CPUStatMonitor
This test module can be used to periodically sample the /proc/stat file
for statistics and report back a list of differences between the
individual samples as well as the raw data.

Can be used to calculate per-cpu and system wide cpu utilization.

Currently the test module samples until interrupted, so it should be run
in the background and stopped with a job.kill(signal.SIGINT) call.

v2:
    * added a comment explaining the interval parameter
    * added a default value for the old_handler variable. In case the
      signal.signal call fails this will be used to avoid a NameError
      exception in the finally block
v3:
    * improved sample parsing based on comment from jtluka, removing the
      useless condition from the loop

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>