add lnst.RecipeCommon.Perf.Measurements package
This is the second part of the refactorization of the PerfAndEvaluate
recipe workflow. In generic terms it introduces a new package that will
store a class hierarchy for various Measurement types and
implementations.

At the base level there is the BaseMeasurement class and module that
defines the interface that all the other classes have to implement. This
interface is understood and relied upon by the
lnst.RecipeCommon.Perf.Recipe class that uses it.

The refactorization includes a move+rename of the IperfMeasurementTool
and TRexMeasurementTool into the new IperfFlowMeasurement and
TRexMeasurement classes/modules. And the addition of the new
StatCPUMeasurement class that uses the CPUStatMonitor test module to
measure cpu utilization.

Finally these changes are added to the BaseEnrtRecipe so that everything
stays working.

v2:
    * fixed typo in IperfFlowMeasurement - the unit for cpu utilization
      returned by Iperf should be "cpu_percent".

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>