refactoring of OvS_DPDK_PvP recipe and related classes
First of all this includes the reimplementation of the TRexMeasurement
module and class into the TRexFlowMeasurement class that implements the
BaseFlowMeasurement API and can be easily plugged into Perf.Recipe as a
measurement. That said, it does have some specific restrictions specific
to TRex:
    * it still requires the trex_dir parameter telling it where to look
      for the TRex application.
    * the measurement is port based but the configuration is flow based.
      Each port currently supports generation of a single flow so that's
      what is expected on the configuration part. However results are
      reported per port (with association to the generated flow). It's
      important to note that while the "tx_rate" statistics represent
      the generated flow, the "rx_rate" statistics only talk about
      received packets regardles of which flow they belong to.

The OvSDPDKPvPRecipe class was updated to work with the redesigned
implementation of the PerfRecipe base class and it's methods for
measurements and reporting results.

The OvSDPDKPvPRecipe now also requests a StatCPUMeasurement measurement
for all the hosts involved in the test.

v3:
    * fixed a trailing white space error

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>