recipes: short_lived_connections: move configuration of multi netperf before loop
This patch avoids repeated configuration of netperf in multi mode when
iterating over nperf_sizes. It's sufficient to configure it before the loop.

Additionally it fixes potential bug when only TCP_CRR is selected for multi
mode. In that case the netperf won't be properly configured because the
configuration remains in TCP_RR branch.

Signed-off-by: Jan Tluka <jtluka@redhat.com>