regression_tests: ovs-dpdk-pvp: add artifical wait time for testpmd
Starting with DPDK 18.11 the startup time increased due to different
memory management, this means we need to add artifical wait times for
the startup to not have our results impacted.

This is a workaround and instead an improvement on the dpdk side should
hopefully happen, but until then we'll need this.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>