lnst.Controller.RunSummaryFormatter: add colourize parameter
The parameter can be used to enable/disable colourization. Currently,
only the PASS/FAIL strings indicating a job success state are supported
for colourization, however this might change later.

Default value is False, this changes the current behaviour where PASS
and FAIL was automatically coloured, but this probably wasn't a good
idea since a normal expectation for a formatter would be to do just the
simplest operation unless specifically configured to something more.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>