lnst.Tests.Iperf: don't fail on stderr messages
Iperf uses stderr to report warnings as well as error messages. However
the test can still run and report results. If there's an actual failure
to execute iperf indicates it by returning a non-zero return value. We
should therefore only report a Fail result on the measurement in case
iperf fails as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>