lnst.Controller.Recipe: init based on class attributes
When initializing a Recipe instance, we're looking for class Param and
HostReq objects to copy and instantiate into the Recipe instance.
However while looping throught the dir() of an instance we can sometimes
attempt to access a "property" or other attributes which aren't valid
during the init method.

The solution is to filter out the type of attributes from the class
itself based on the type that we want to work with. This also simplifies
the code a little.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>