RecipeCommon.Perf.Evaluators.BaselineCPUAverageEvaluator: add filter logic
Adding a new parameter to the CPU evaluator - evaluation_filter. This
Parameter expects a dictionary where:
key - represents a hostid as accessed through Host.hostid property
value - list of strings representing cpu names

This dictionary will be used to pre-filter the results to evaluate. This
can be helpful when dealing with machines that have many cores and only
some are strictly used during the performance test. All the other cores
might be idling or the scheduler assigneds them tasks differently
between runs which can lead to confusing evaluations.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>