Perf.Measurements.BaseMeasurement: add recipe_conf attribute
This attribute should contain an object desribing the complete
configuration of the recipe that is relevant to the perf measurement.
This carries mostly informational value for reporting and evaluation.

In our use-case it will be used during evaluation when searching for
baselines in the database to distinguish between different measurements
generated by the same recipe, e.g. ipv4 vs ipv6.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>