ENRT.BaseEnrtRecipe: remove unused parameters
These were never used and seem to be useless so I'm removing them. I'm
also changing the perf_reverse parameter to have a default value of
False since it makes the logic of using it in the recipe easier (only
check it's value instead of also checking if it's defined).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>