ENRT.BaseEnrtRecipe.EnrtConfiguration: add params attribute
EnrtConfiguration object should contain information about the test wide
configuration, that is relevant during the entire execution of the
recipe. Most of this is directly represented by the values of recipe
parameters.

I'm adding the params attribute to the EnrtConfiguration class as a way
to have access to a copy of these values during parts of the test that
don't have access to the original Recipe object.

The most logical place where this will be used is PerfResult evaluation.

However I still consider this just an experimental approach as I think
it might not work very well for all the enrt recipes that we have and a
larger code refactoring of this code might be required.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>