lnst.RecipeCommon.Perf.Evaluators: generalize Baseline evaluators
Based on experiences with the two Baseline evaluators (CPU and Flow
based) they share some common mechanisms and design which should be
generalized into a base class to reduce code duplication.

This commit introduces the BaselineEvaluator class which defines the
common methods and the general mechanism of evaluating results by
comparing against a baseline. It also provides default implementations
for these methods which means that the class can be used on it's own,
though all evaluation results will be False (failures).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>