lnst.Controller.Machine: ensure Job{Start, Finish}Result ordering
Sometimes the job finished message could arrive from the slave before
the job started message which would lead to an inverse ordering that
couldn't be shortened by the ResultFormatter into a single line. It's
also a little confusing to see that a job finished before it was
started.

This fixes the issue by pre-emptively adding a successful
JobStartedResult to the current_run result log and changing it's
"success" attribute once the result from the slave is known.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>