add __repr__ and __str__ implementations of several objects
__repr__ methods return a string that can be put back to python to
interpret it and create an "equal/identical" object instance. This can
be quite useful when printing logs - the method makes the objects self
describe themselves without any specific format.

The exception is that some of these can't be cleanly interpreted because
they require references to some internal objects that would be clumsy to
also provide... In those cases, the __str__ method is defined instead
which only provides some informal description of the object but in a
similar format.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>