lnst.Devices.RemoteDevice: add _id property
The property contains the name of the attribute that the class instance
was first assigned to when created on a Namespace instance. Example:

host1.my_vlan_device = VlanDevice(...)

The _id attribute in this case is "my_vlan_device".

The attribute starts with "_" to indicate that it's for internal LNST
code use. Specifically it will be used for logging messages involving
the device.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>