lnst.Controller.RunSummaryFormatter: change default source string
The default source string used by the summary formatter was "", so when
a new type of result class was created this part of the summary would be
left blank. This is kind of hard to notice and the developer might
forget that behaviour for this new result class should be defined in the
summary formatter as well.

Changing the default to the class name instead at least makes sure that
something shows up and if it looks bad it will be more noticeable.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>