lnst.Controller.Job: __str__ method format of _what description
Using the repr() of the _what object instance should lead to a more
comprehensive description of what the job is, especially for jobs that
execute a test module.

I also moved this to the end of the generated string since this is the
field with most length variance, this way it makes the line formatting
more consistent when multiple jobs are printed under each other.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>