lnst.Controller.Machine: add device manipulation methods
Moving the device related rpc_calls from the RemoteDevice class into the
Machine class which already provides abstractions for all the other rpc
calls that we do. It's also a good place to organize the creation of
Device related event results and add them to the current run log of the
recipe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>