adjust logging levels
Downgraded most of the info level logging messages to the debug level as
they're often times not that helpful with understanding what the recipe
is testing.

At the same time I set the default result level to IMPORTANT for every
Job run on a host (through the host.run(), or host.prepare_job()
methods). This changes the approach for the testers:
- the original approach was to assume that jobs run by the testers were
unimportant unless specifically marked as important
- the new approach flips this assumption - all jobs by the testers are
important, unless specifically marked as not.

This should hopefully make it more natural for a tester to understand
the logs and result summary of a recipe.

I went ahead and adjusted the result levels of some smaller jobs in the
StatCPUMeasurement and BaseEnrtRecipe code to keep the original
behaviour.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>