lnst.Devices.Device: refactor nl_sync methods
Created a new _ipr_wrapper method that wraps any kind of IPRoute()
object use that we need in the Device classes.

Renamed the _nl_sync method to _nl_link_sync to indicate that this
method handles IPRoute().link methods specifically. This differentiates
it a little better from the related IPRoute().addr methods used for
manipulating IPs of a device.

Centralized error handling into the new _ipr_wrapper method to improve
consistency of the logs.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>