lnst.Controller.RecipeResults: reduce the level of DeviceConfigResults
IMPORTANT is the default result level that gets printed by the
ResultSummaryFormatter. Seeing all the individual config actions can
make it quite verbose and hard to see the actual tests. Reducing the
level to NORMAL will clear it out from the summary by default, while
retaining some importance and the config actions will still be visible
in the full logs of the recipe execution.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>