add lnst.Recipes.ENRT.ConfigMixins package
The ENRT.ConfigMixins package will hold any config related modules that
are common and can be implemented in a mixin fashion to significantly
reduce code duplication in our recipes.

For a start I'm adding the BaseSubConfigMixin class which defines the
base interface of any sub configuration related classes. One such class
will the OffloadSubConfigMixin that will take over the offload
configuration instead of having it directly in the BaseEnrtRecipe class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>