lnst.Recipes.ENRT.BaseEnrtRecipe: refactor the ping/perf test loops
Refactoring the ping and perf test loops into smaller methods and
refactoring the generate_{ping, perf}_configuration methods to use
endpoint generator methods instead of accessing a static attribute of
the recipe config object.

This makes the tests more flexible as each recipe can define its own
endpoints, as many of them as is relevant and also as many in parallel
as is required. It can also implement its own logic and generate the
endpoints dynamically.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>