lnst.Recipes.ENRT.BaseEnrtRecipe: remove EnrtConfiguration attributes
The configuration object should now store arbitrary data that each
derived recipe decides to put in it instead of providing a strict api
interface.

The object will get automatically created by the default
test_wide_configuration method, however testers are free not to use this
functionality. It's current implementation is there to enable
collaborative inheritance but if you override it the recipe should still
work fine.

Later on it might make sense to introduce a factory method that creates
an EnrtConfiguration derived object that is defined by a more specific
class and enforces an api, but I don't think we need that for now.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>