lnst.Recipes.ENRT.SimplePerfRecipe: add test wide description generation
Updating the SimplePerfRecipe class to use an EnrtConfiguration object
created by the base class and tracking a new attribute
(test_wide_devices) that is relevant to the SimplePerfRecipe class with
the configured devices.

Adding an implementation of the generate_test_wide_description method
that uses this new attribute to add a description of what was
configured.

This also implements the test_wide_configuration method in a way that
enables collaborative inheritance to work properly so that we can extend
the recipe functionality by simply adding a parent class that adds some
generic test wide configuration.

This way the EnrtConfiguration object is created by the BaseEnrtRecipe
class and only extended by the individual test_wide_configuration
methods.

---
v2:
* renamed parent to desc in generate_test_wide_description

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>