add ConfigMixins.OffloadSubConfigMixin module, enable for SimplePerfRecipe
The OffloadSubConfigMixin is an EnrtBaseRecipe subconfiguration mixin
class implementing the offload subconfiguration looping. This includes
generating, applying, removing and describing the subconfiguration.

It also includes overriding the default perf test flow generation loop
so that certain test configurations are ignored when certain offloads
are configured.

Enabling this mixin for the SimplePerfRecipe means simply adding it as a
base class and defining the required property 'offload_nics' indicating
which devices should the offload configuration be done for.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>