add lnst.Recipes.ENRT.ConfigMixins.BaseHWConfigMixin hierarchy
The BaseHWConfigMixin class hierarchy implements various device
configuration scripts related to hw devices that we almost always do in
the same order and on the same set of devices. It includes:
* mtu configuration
* device interrupt cpu pinning
* coalescing configuration
* QDisc configuration in case parallel stream tests are requested

Each configuration is implemented in its own mixin class inheriting from
the BaseHWConfigMixin class that defines the interface. Additionally the
CommonHWConfigMixin class inherits from all 4 config mixins and
implements a collaborative inheritance version of the
test_wide_configuration, deconfiguration and description methods to be
added to the BaseEnrtRecipe derived classes.

---
v2:
* renamed parent to desc in description generation methods

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>