lnst.Recipes.ENRT.SimplePerfRecipe: implement wait_tentative_ips
After configuring IPv6 addresses and setting a device state to UP, it's
required to wait for the ip address to leave the tentative state during
which the ip address isn't bindable by network applications.

The tentative state is used to indicate that the ip address is
undergoing duplicate address detection on the network.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>