lnst.RecipeCommon.Perf.Evaluators.BaselineEvaluator: fix return value type
The return value of compare_result_with_baseline should be a list of
strings that describe the comparison. The result gets used in the
evaluate_group_results method which joins all the lists together.

Returning a string here would still work becuase it can be iterated but
it would result in incorrect formatting of the result.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>