CommonHWConfigMixin: refactoring into a SubConfig class
Device interrupts need to be configured every time an "up/down" cycle
happens for a device, as the ENRT recipes are designed around a
sub_configuration loop that often times involves such configuration, the
order of CommonHWConfigMixin needs to be adjusted.

After additional discussion with Ivan Vecera I decided that it makes
quite a bit of sense to look at common hw configuration as just another
SubConfigMixin class that should be called as the final sub config class
in the BaseENRT sub config loop.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>