lnst.Devices.Device: add DeviceFeatureNotSupported for adaptive coalescing
For now throwing an exception when trying to read/write coalescing
settings to a device that doesn't support them is still the best option.

However, there's a need to easily distinguish and ignore this error in
operations that try to access all properties of a Device class instance,
e.g. enable_readonly_cache. In such cases the tester probably doesn't
care that coalescing isn't supported and has no option to skip it for
the cache generation operation leading to the recipe failing with an
exception.

Handling device capabilities dynamically is a larger feature that might
be hard to implement for now so instead a special
DeviceFeatureNotSupported exception class is added that can be safely
ignored when performing these "bulk"/"iterative" operations.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>