lnst.RecipeCommon.Perf.Results: fix for zero division errors
Instread of throwing an exception we should simply return a float
infinity value so that it propagates all the way to the user.

Same exception could also be raised by the BaseFlowMeasurement "report"
methods while calculating the percentage of the standard deviation out
of the average value.

A helper method was utilized to calculate the percentage for all of the
measured values.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>