lnst.Slave.InterfaceManager: redesign rescan_devices
The rescan_devices method historically used the NetUtils.scan_netdevs
method to externally dump the current configuration state of network
devices.

This introduced several race condition issues due to having 2
sockets/communication channels for the same information. The main
problem was that messages received on the main
InterfaceManager._nl_socket should be deprecated when a rescan_devices is
called, however in some situations some messages may already be being
processed and saved in a temporary list when a rescan is called.

This could lead to removing a device during the rescan_devices call,
then adding it back when handling of nl_socket messages resumes, leading
to all sorts of double remove issues.

The redesign to address this issue is to remove the second channel
created by the scan_netdevs method. The direct netlink socket connection
which we use to listen for notifications can also be used to directly
request the dump of the configuration state and add it to the end of the
FIFO message queue in the correct order, resolving any possible race
conditions.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>