lnst.Controller.Machine: add device method result after the method call
Adding the result before calling the method leads to printing a "live
log" line that says that the action is a "PASS", but the method wasn't
called yet.

Changing the result to False after the method call is therefore only
visible in the Result Summary logs, which can lead to confusing
PASS/FAIL situations when a tester sees the PASS in live logs, but FAIL
in summary for the same action.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>