lnst.Recipes.ENRT: add VhostNetPvPRecipe
Inherit common functionality from BasePvPRecipe

Use a bridge based forwarding plane both in guest and host.
In the future it may be worth adding more forwarding planes (tc,
xdp, etc)

In this case the traffic flows from one trex interface to the other
making the bridge in the guest do the switching, so the destination
IP address of the TRex flows is the generator's.

Use taskset to pin the vhost-net kernel threads.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>