lnst.Recipes.ENRT: add pause frame disabling HWConfigMixin
We've historically been disabling pause frames at the start of device
configuration by default. This functionality was skipped when
implementing lnst-next since it shouldn't be the default behaviour of
Device configuration classes and was forgotten when implementing the
ENRT recipes.

This commit adds a new hw config mixin and adds it to the list of
CommonHWSubConfigMixins.

We also enable the use of the mixin by the SimpleNetworkRecipe class by
defining the list of devices to disable pause frames on.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>

---
v2: fixed confirmation command which doesn't accept additional
    parameters