TRexFlowMeasurement: attach server to parametrized CPU cores
The TRexFlowMeasurement object now needs to be initialized with a list
of cores to use for the TRexServer process instead of hard coded values.

This requires a change to the OvSDPDKPvPRecipe which previously removed
a parameter that was intended to be used for this but skipped.

The same parameter needs to be added to the VhostNetPvPRecipe which also
uses the TRexFlowMeasurement.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>