Sphinx based documentation
Creating the docs/ directory where we should work on creating
documentation. The standard for Python projects is to use "Sphinx" to
generate various formats (e.g. pdf, html or others) of documentation
from ReStructuredText (rst) source. The source is either stored as
".rst" standalone files or as docstrings with Python source code.

This is the first version that should cover some of the major topics and
exported APIs available for the tester. It is by no means complete yet
but should be enough to get a feel for how the overall structure should
look like.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2:
    * add master_doc to conf.py
    * renamed docs source files for consistency
    * fixed SlavePoolManager reference
    * added 'none' to ascii art code-block
    * simplified wording for sub configuration explanation
    * swapped sentences for net_perf_tool and cpu_perf_tool parameters
    * updated docs for describe_test_wide_configuration method

v3:
    * fixed --requirements -> --requirement typo
    * rebase on current master

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>