lnst.Devices.Device: update the address flags if they have changed
When IPv6 address is added to an interface there are actually two
RTM_NEWADDR netlink messages sent. The second message is sent
when DAD is complete and tentative flag is removed.

For the second message associated with the same address we have to
check if the flags have changed and update them in the Device instance
to be in sync.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>