remove NetworkManager support
LNST no longer configures the network stack through NetworkManager
and we have no plans to enable it in the future.

This patch removes the configuration entries from lnst-slave.conf config file
and the unused RPC methods from NetTestSlave and Machine code.

I kept the warning log entry if NetworkManager is detected on the slave
machine, as this is still important information.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>