Tests.Ping: update failure logic
This patch modifies the code to leave the pass/fail decision to a user of the
test module.

The module will now return with failure only if exit code is higher than 1.
Exit code 1 is used to report that ping did not receive any reply packets
at all or fewer packets than expected are received. Ping exits with exit
code 2 on other errors.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>