lnst.Recipes.ENRT: add PingMixins
These mixins will be used to override the default Ping evaluator of the
BaseEnrtRecipe class.

The module provides VlanPingEvaluator mixin that will be used in all
vlan-like (e.g. VLAN, VXLAN) tests where we try to ping from a vlan to
another one that is meant to be unreachable.

v3:
 * remove PingConditionMixin completely

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>