lnst.Common.ConnectionHandler: clean up closed connections
Before calling select on provided connection/socket objects we need to
check if any of them may have been closed and need to be removed first.

The main reason for this is that a closed multiprocessing.Connection
object may be present in the list which will result in the following
select call returning with an exception. This exception doesn't indicate
which of the file objects causes the issue and selecting for "error"
file descriptors also doesn't help. As such we need to check this before
calling select.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>