Devices.OvsBridge: use ovs-vsctl list to get ovs bridge ports
This is an additional patch that removes incorrect reference to
internal ports. The property internal_ports is removed and
ports with internal interfaces are now returned in property ports.

The patch also fixes an issue with OvS 2.13 that changed
the output format of 'ovs-vsctl show' that was used to get
information about the ovs bridge ports. To be more future proof
the json output of 'ovs-vsctl list ports' is used instead.

Since tunnels are also ports, the property tunnels reuses the ports
property and filters out the relevant ports.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>