Recipes.ENRT.BaseEnrtRecipe: skip cross-vlan ping for IPv6
With the current implementation of ping6 there's no way to test
a ping from one vlan to a different one in a similar way to IPv4.

When ping6 -I <ip> makes a bind(sockfd, &ipv6address) the bind is only
a hint so it tries to use the address but if it fails it uses the other
address from the vlan interface.

We had the same issue in legacy LNST and conclusion was to disable
this test for IPv6.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>