Recipes.ENRT: remove cross-vlan ping test from Vlans recipes
The current implementation of the vlan isolation test is not accurate.
In particular for the VlansRecipe, VlansOverBondRecipe and VlansOverTeamRecipe,
where under certain conditions (e.g. rp_filter=2 or IPv6) the vlans are
expected to be reachable.

Before we find a better way to test this I removed the relevant PingEndpoints
in these tests.

I kept the similar test in virtual tests because in these topologies
the test makes sense. The virtual guests are isolated at the hypervisor
level and must not be be able to access the other network.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>